refactor: Chu, Han 구현체 Team 단일 클래스로 변경
